Renamed file and absorbed functions
