+++ tt_metal/hw/inc/api/debug/timing_perturbation.h
@@ -1,0 +1,56 @@
+// SPDX-FileCopyrightText: © 2026 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+// This file contains common timing perturbation utilities for debugging
+
+#if defined(COMPILE_FOR_TRISC) && defined(DEBUG_PRINT_ENABLED) && !defined(FORCE_DPRINT_OFF)
+
+namespace tt::compute::common {
+
+/**
+ * @brief Helper to inline variable number of nops/TTI_NOPs
+ *
+ * @tparam num_nops: number of nops to insert
+ * @tparam nop_type: 0 for accelerator TTI_NOP, 1 for RISC-V nop
+ */
+template <const int num_nops, const int nop_type>
+inline void add_nops() {
+    for (int i = 0; i < num_nops; i++) {
+        if constexpr (nop_type) {
+            asm volatile("nop");
+        } else {
+            TTI_NOP;
+        }
+    }
+}
+
+/**
+ * @brief Uses template parameters to insert variable compute nops for timing perturbation
+ *
+ * @tparam num_unpack_nops: number of unpack nops to insert
+ * @tparam num_math_nops: number of math nops to insert
+ * @tparam num_pack_nops: number of pack nops to insert
+ * @tparam nop_type: 0 for accelerator TTI_NOP, 1 for RISC-V nop
+ */
+template <const int num_unpack_nops, const int num_math_nops, const int num_pack_nops, const int nop_type>
+inline void add_compute_nops() {
+    DPRINT << "NOP Type: " << (uint32_t)nop_type << " Unpack NOPs: " << (uint32_t)num_unpack_nops
+           << " Math NOPs: " << (uint32_t)num_math_nops << " Pack NOPs: " << (uint32_t)num_pack_nops << ENDL();
+    if constexpr (num_unpack_nops) {
+        UNPACK((add_nops<num_unpack_nops, nop_type>()));
+    }
+
+    if constexpr (num_math_nops) {
+        MATH((add_nops<num_math_nops, nop_type>()));
+    }
+
+    if constexpr (num_pack_nops) {
+        PACK((add_nops<num_pack_nops, nop_type>()));
+    }
+}
+}  // namespace tt::compute::common
+
+#endif  // defined(COMPILE_FOR_TRISC) && defined(DEBUG_PRINT_ENABLED) && !defined(FORCE_DPRINT_OFF)
